Use simpler roundoff rules
I used to like rounding based on significant figures,
until I read [1], which highlights some problems with it.

It is better to report 2 numbers: the value and percentage or stdev.
RAGE does this most of the time, except when stdev is <5% in which case
it reports only 1 value.

The value we report can be rounded, but the rounding shouldn't be based
on the standard deviation: a simple rule that rounds to a fixed number
of significant digits is better. The uncertainty is already expressed in
the 2nd number, so there is no need to introduce additional errors by
rounding the 1st number.

NIST [2] also recommends reporting 2 numbers.

Also use the unrounded numbers for further calculations, only round off
what gets displayed to the user.

[1] https://www.av8n.com/physics/uncertainty.htm
[2] https://physics.nist.gov/cuu/Uncertainty/examples.html

Signed-off-by: Edwin Török <edvin.torok@citrix.com>